API

Dilutions.normalize

Dilutions.normalize

Dilutions.standards

Dilutions.normalize — Method

normalize(source_concentration, target_concentration, target_volume)
normalize(df::DataFrame)
source_concentration: concentration of stock
target_concentration: needed concentration
target_volume: needed volume

The values can be either numeric or with units from Unitful.
If units are used, the result will be in the unit of `target_volume`.
Values can also be given as keyword arguments.
For a DataFrame, it must have columns matching `r"[source|stock]_con", r"target_con", r"target_vol"`

Result: (;source_volume, solvent_volume)n

Examples:
``` julia
normalize(5000,200,1000) == (source_volume = 40.0, solvent_volume = 960.0)
normalize(5000u"µmol/L",200u"µmol/L", 1000u"µL") == (source_volume = 40.0u"μL", solvent_volume = 960.0u"μL")
normalize(5u"mmol/L",200u"µmol/L", 1u"mL") == (source_volume = 0.04u"mL", solvent_volume = 0.96u"mL")
```

Dilutions.standards — Method

standards(source_concentration, target_concentrations, target_volume)

source_concentration: concetration of stock
target_concentrations: vector of target concentrations
target_volume: desired volume of each standard

result:
DataFrame with columns: source_concentration  target_concentration  target_volume  source_volume  solvent_volume
where the last 2 are instructions in how to make the standard solutions.

Currently this is very simple using only the given stock concentration.
A more realistic version will allows for using intermediate dilutions for the lower concentrations.